Move common ft includes to FtIncludes.h.
